Add createSlice & create builder cases
