+++ src/redux/configureStore.js
@@ -1,13 +1,16 @@
-import { applyMiddleware, combineReducers, legacy_createStore as createStore } from 'redux';
+import { applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
-import userReducer from './users/registerSlice';
 import { configureStore } from '@reduxjs/toolkit';
-import registerReducer from './users/registerSlice';
+import registerReducer from './users/userSlice';
+import { createLogger } from 'redux-logger';
+
+const logger = createLogger();
 
 const store = configureStore({
   reducer: {
     register: registerReducer,
-  }
+  },
+  middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(logger),
 })
 
 export default store;
